feat: :art: Refactorizacion de estructura de archivo de componentes
