--- src/constants/chatMessages.js
@@ -1,30 +1,0 @@
-// constants/chatMessages.js
-export const DEFAULT_MESSAGES = {
-    WELCOME: {
-      id: 'welcome',
-      avatar: 'B',
-      message: '¡Hola! Soy Iza, y estoy aquí para ayudarte a descubrir cómo Campuslands puede transformar tu futuro 😊.',
-      isAI: true
-    },
-    AGE_FORM: {
-      id: 'age-form',
-      avatar: 'C',
-      message: '🎉 Nos encantaría conocerte un poco más\n\n¿Qué edad tienes?',
-      isAI: true,
-      type: 'age-form'
-    },
-    AVAILABILITY_FORM: {
-      id: 'availability-form',
-      avatar: 'C',
-      message: '¿Te animas a dedicar 8 horas diarias de manera presencial durante 10 meses? 😊',
-      isAI: true,
-      type: 'availability-form'
-    },
-    NEXT_STEPS: {
-      id: 'next-steps',
-      avatar: 'C',
-      message: '🚀 ¡Excelente! ¡Gracias por tu respuesta! ¿Hay algo específico en lo que necesites ayuda? 🌟',      isAI: true,
-      type: 'next-steps-form'
-    }
-
-  };
