+++ src/services/apiConfig.js
@@ -1,5 +1,4 @@
 const API_BASE_URL = import.meta.env.VITE_API_BASE_URL || "https://chatcampuslands.com:8443/chatbot/";
-const API_WEBSOCKET_URL = import.meta.env.VITE_API_WEBSOCKET_URL || "wss://chatcampuslands.com:8443/chatbot/chat";
 
 // console.log("Variables de entorno: ",import.meta.env.VITE_API_BASE_URL, " y ",import.meta.env.VITE_API_WEBSOCKET_URL);
 
@@ -11,4 +10,3 @@ export const endpoints = {
   availability: `${API_BASE_URL}user/availability`,
 };
 
-export default API_WEBSOCKET_URL;
